backend/requirements.txt
alembic==1.17.2
asyncpg==0.31.0
bcrypt==5.0.0
cryptography==46.0.3
dishka==1.7.2
fastapi==0.128.0
pydantic==2.12.5
pydantic-settings==2.12.0
PyJWT==2.10.1
python-dotenv==1.2.1
python-multipart==0.0.21
redis==7.1.0
SQLAlchemy==2.0.45
uvicorn==0.40.0
taskiq==0.12.1
inflect==7.5.0
taskiq-aio-pika==0.5.0
aiosmtplib==5.1.0
pyotp==2.9.0
black
isort
flake8